requirements.txt
dj-database-url==0.5.0
Django==2.0.5
django-heroku==0.3.1
djangorestframework==3.8.2
gunicorn==19.8.1
psycopg2==2.7.4
pytz==2018.4
whitenoise==3.3.1
faker==0.8.15
django-cors-headers==2.2.0
django-filter==1.1.0
django-rest-auth==0.9.3
